refactor and move web-component-service.js to main src folder.
